getting genbank files and parsing them
I updated the api tool to be able to get genbank files for all phages that infect a given host. I then created a parser that searches through the genbank file for all protein sequences and saves them in a fasta file. I used this tool to generate fasta files of phage protein sequences and host protein sequences.